extract request util
